requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
flask==3.0.2
greenlet==2.0.2
grpcio==1.60.1
grpcio-tools==1.60.1
idna==3.6
isodate==0.6.1
keyboard==0.13.5
lxml==4.9.3
mysql-connector-python==8.3.0
platformdirs==4.2.0
protobuf==4.25.2
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.1
python-socketio==5.11.1
pytz==2023.3
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
spyne==2.14.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.2.0
websocket-client==1.7.0
zeep==4.2.1